round 5 of esm migration, for src/app.js and its importers
